refactor(account)!: move methods from AccountBase to AccountMemory
BREAKING CHANGE: `AccountMemory` requires `networkId` in `signTransaction`

BREAKING CHANGE: `AccountBase` simplified
- `networkId` removed
- `getNetworkId` method removed
- `signTransaction`, `signMessage` made abstract
