+++ src/account/Memory.ts
@@ -15,10 +15,17 @@
  *  PERFORMANCE OF THIS SOFTWARE.
  */
 import AccountBase from './Base';
-import { generateKeyPairFromSecret, sign, generateKeyPair } from '../utils/crypto';
+import {
+  generateKeyPairFromSecret, sign, generateKeyPair, hash, messageToHash,
+} from '../utils/crypto';
 import { isHex } from '../utils/string';
 import { ArgumentError } from '../utils/errors';
-import { encode, Encoded, Encoding } from '../utils/encoder';
+import {
+  decode, encode, Encoded, Encoding,
+} from '../utils/encoder';
+import { concatBuffers } from '../utils/other';
+import { buildTx } from '../tx/builder';
+import { Tag } from '../tx/builder/constants';
 
 const secretKeys = new WeakMap();
 
@@ -32,7 +39,7 @@ export default class AccountMemory extends AccountBase {
    * @param secretKey - Secret key
    */
   constructor(secretKey: string | Uint8Array) {
-    super({});
+    super();
     if (typeof secretKey === 'string' && !isHex(secretKey)) {
       throw new ArgumentError('secretKey', 'hex string', secretKey);
     }
@@ -54,7 +61,28 @@ export default class AccountMemory extends AccountBase {
     return new AccountMemory(generateKeyPair().secretKey);
   }
 
-  async sign(data: string | Uint8Array): Promise<Uint8Array> {
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  override async sign(data: string | Uint8Array, options?: any): Promise<Uint8Array> {
     return sign(data, secretKeys.get(this));
+  }
+
+  override async signTransaction(
+    tx: Encoded.Transaction,
+    { innerTx, networkId, ...options }: { innerTx?: boolean; networkId?: string } = {},
+  ): Promise<Encoded.Transaction> {
+    if (networkId == null) {
+      throw new ArgumentError('networkId', 'provided', networkId);
+    }
+    const prefixes = [networkId];
+    if (innerTx === true) prefixes.push('inner_tx');
+    const rlpBinaryTx = decode(tx);
+    const txWithNetworkId = concatBuffers([Buffer.from(prefixes.join('-')), hash(rlpBinaryTx)]);
+
+    const signatures = [await this.sign(txWithNetworkId, options)];
+    return buildTx({ encodedTx: rlpBinaryTx, signatures }, Tag.SignedTx).tx;
+  }
+
+  override async signMessage(message: string, options?: any): Promise<Uint8Array> {
+    return this.sign(messageToHash(message), options);
   }
 }
